requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
eventlet==0.33.3
Flask==2.2.5
Flask-Cors==4.0.0
Flask-SocketIO==5.3.4
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
pymongo==4.4.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.31.0
six==1.16.0
tqdm==4.65.0
urllib3==2.0.3
Werkzeug==2.2.3
yarl==1.9.2
